refactor(Web Search): 精炼 Provider 与异常映射
将 Provider 响应标准化工具收回为私有模块，内联工具执行异常映射，并补充核心模型职责说明，减少无意义跳转且保持原有错误语义不变。
