+++ services/wisepen-chat-service/src/chat/application/tools/search_tools/web_search/services/providers/_utils.py
