+++ services/wisepen-chat-service/src/chat/application/tools/search_tools/web_search/services/providers/platform_default.py
@@ -19,7 +19,7 @@ from .core.models import (
     SearchPreview,
 )
 from .core.protocols import ProviderSearcher
-from .normalization import (
+from ._utils import (
     as_dict_tuple,
     as_str,
     as_str_or_none,
